requirements.txt
anyio==3.6.1
APScheduler==3.9.1
cachetools==5.2.0
certifi==2022.9.24
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
packaging==21.3
pdf2image==1.16.0
Pillow==9.2.0
pyparsing==3.0.9
pytesseract==0.3.10
python-telegram-bot==20.0a4
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
telegram==0.0.1
tornado==6.2
tzdata==2022.4
tzlocal==4.2
